+++ gs/src/gdevdjet.c
@@ -104,14 +104,14 @@
  * inexact paperlength which is set to 117 10ths.
  * Somebody should check for letter sized paper. I left it at 0.07".
  */
-#define DESKJET_MARGINS_LETTER  0.2, 0.45, 0.3, 0.05
-#define DESKJET_MARGINS_A4	0.125, 0.5, 0.143, 0.09
+#define DESKJET_MARGINS_LETTER  (float)0.2, (float)0.45, (float)0.3, (float)0.05
+#define DESKJET_MARGINS_A4	(float)0.125, (float)0.5, (float)0.143, (float)0.09
 /* Similar margins for the LaserJet. */
 /* These are defined in the PCL 5 Technical Reference Manual. */
 /* Note that for PCL 5 printers, we get the printer to translate the */
 /* coordinate system: the margins only define the unprintable area. */
-#define LASERJET_MARGINS_A4	0.167, 0.167, 0.167, 0.167
-#define LASERJET_MARGINS_LETTER	0.167, 0.167, 0.167, 0.167
+#define LASERJET_MARGINS_A4	(float)0.167, (float)0.167, (float)0.167, (float)0.167
+#define LASERJET_MARGINS_LETTER	(float)0.167, (float)0.167, (float)0.167, (float)0.167
 
 /* See gdevdljm.h for the definitions of the PCL_ features. */
